... figured out problem, changing deliteralization
... to do something like cloned-value info.  This basically just sets the
`literal?` and `literal` slots to `f`, but keeps the `slots` slot intact.  That
one is expected to still contain something during `infer-value`, which is
triggered when following the custom-inlining of `call-effect`.

Since we don't want to cache on the whole signature again, also have to reset
the interval to the class-interval.