... simplified dependency tracking during inline-propagation
- While the body is propagated without dependencies, we store a link to the
top-level dependencies vars, and record inlining there after we 1.
- decided to inline anyways
- actually inferred non-object info somewhere

Note that this means that we only create the dependency if we relied on anything
specific, and this may prevent the calling word from being recompiled even if
the lower level is optimized...
One compromise could be to only check for non-trivial info on non-method words,
and register methods as dependencies in any case.

Also use the inline-propagation flag now.
Also add some tests already for dispatch inlining.