compiler.tree.propagation: Correctly identify escaping values
Rationale: Only pointers can escape.  This is equivalent to the slots in the
value-info states.  These are assigned `lazy-info`s.  Their values are distinct
from the ones that are propagated during regular "value-level" propagation.
When escaping, collect the closure of reachable lazy-infos.  Invalidate all
infos of the virtual values accordingly.