compiler.tree.propagation: Remove explicit de-literalization
Is now part of init-value-info procedure.
The biggest change is that during init-value-info, the literal and flag is kept
if a literal has been modified.  Full de-literalization would kill that flag,
messing up parts of the compiler that depend on it.  The correct semantics would
be something like "literal-and-read-only-slots-not-modified", but that seems
complex to implement.