compiler.tree.propagation: Allow lazy-non-lazy-info unions
Concerns propagation at loop headers, specifically merging via counter
generalization.  Note that this in itself is not correct yet.

Problem: Virtuals generalization is done separately from regular
generalization.  However, this is not enough/wrong:  If a new virtual value is
incorporated into generalization, virtual counter generalization will not catch
this.  Currently we treat a merge at the loop header like a regular #phi merge,
but that is not correct concerning the semantics.