Automatically define singletons for words for dispatching
Uses implicit "<vocab>.singletons" vocab to define a predicate class equivalent
to a singleton class.

Whenever using scan-class, this mechanism is plugged in to JIT-define this
helper class instead of throwing an error.

The helper predicate class is defined on `word`, except for when the word itself
denotes a class, in which case it is defined on `class`.

... do it in scan-class instead

also define the predicate on `class` if the word denotes a class