types: Unify using pred/succ "combinators"
Instantiate all possible series members in type schemes as variables, solve
using explicit pair-wise successor/predecessor constraints.  Drop is handled by
using some sort of "splice" combinator which interacts with the S and
P (successor and predecessor, respectively) terms in the constraints.  The idea
is to force a "rebuild" of the types in the resulting scheme to account for both
directions of inference.