types: Use fi in combinator base
Turns out that for linearity purposes, we cannot recover `if` easily.  Thus,
make `if` a base combinator instead, from which it is easy to recover `k`.

Also implements a complex type for representing that either type can be
returned, but not both.  Thus, as long as we don't know anything about the
associated condition, we would have to interpret this as sum.  This is in
contrast to the input type matching, which we require to be possible for both
types.