chr.factor.composition: Deactivate some disjoint assumptions
The idea behind these was that if we had an unknown call or effect type in some
branch, they may not be merged during a union effect, i.e. these values
automatically serve as "parameters" for the branch effect types.

Deactivated them for now, as they are not active anywhere, and they might impair
reasoning with loops...