chr.factor.composition: Merge Phi effect vars, but not in some cases
Reasoning becomes complicated if explicit `{ Phi out a b }` Predicates
are used, because all other sub-solvers must support these explicitly.

However, simply unifying comes with the problem of possibly introducing wrong
equality constraints, if the definition of an output stack element can be
ambiguous depending on the branch taken.  This happens e.g. when joining
swap/non-swapped branches, or branches with a lingering dup.

These cases are now not even tried for phi unification.  The idea is that these
are cases with unresolved data flow, i.e. continuation-level fragments, which
can not be assigned pure value types anyway.