chr.factor.composition: Don't automatically throw out duplicate preds
During compositional reasoning, they are redundant, so we only collect them if
we don't already have them, and they will be collected anyway if left over.

During union reasoning, we need to know if we have exactly two instances of a
predicate, because that means that we have this restrictions in both branches,
and can keep it alive.