chr.factor: Support conditional macro expansion (somewhat)
Macro expansions with conditionals are inserted as Xor instance predicates
with CallEffect predicates.  Regular effect composition avoids this, as all Xor
types are usually distributed through the top-level.  However, for delayed
expansion, like in macros and probably unknown instance checks as well, these
are necessary, since we don't know beforehand that there will be an Xor type.

Thus, this predicate combination is caught during intra-effect reasoning,
usually triggered by the re-inference caused by a macro expansion type becoming
known.

This is then searched for in any type to be completed, which is re-inferred to
become an Xor type instead.

Problem: This does not work for expansion into nested Xor types right now.
Approach: Do the check for CallXorEffect recursively, rebuilding the
corresponding Xor Type recursively.

Note: This is can already infer more code than the current Factor spec.  It only
really is necessary for handling incompletely expanded macros, which are not
supported by Factor.  Thus, as an optimization, the quotation decomposer should
probably be modified to make sure any expansions are caught as early as possible.