terms: Add limited set matching capability
This works for solving problems like

HS{ 1 2 { ?foo } { 5 } } ==? HS{ 2 1 { 5 } { 42 } }

but not for

HS{ 1 2 ?foo } ==? HS{ 1 2 3 }

The reason is that we pre-check on the element classes and their cardinalities,
and then perform a subset-by-subset matching with backtracking on the entire
remaining solver continuation.
This is a naive DFS in the search tree, so has extremely high complexity!

One possible optimization: postpone the matchset sub-problems as long as
possible.  This should be possible since the solver result must be independent
of the destructuring order, so we can do the "cheaper" checks first.