chr.factor: Nested Inference for macros, rework phi stuff
Inference for macroexpansions is now done during composition inference,
by suspending the current inference, inferring the expanded quot type, and
returning.

The criteria for merging Xor effects have changed slightly.  It is now not
required for one Decider and Discriminator be in input and output, or vice
versa.  Rather, two effects will be kept disjoint if there is an observable
decider value, and any relations which could cause a "significant" difference
between values in the branches.  This is regardless of input or output
positions.

Also got rid of the Params predicate.  Effect parameters are now used
exclusively for holding the scope during alpha-renaming.

Isomorphic checks have been expanded to be agnostic to predicate order
in effect bodies, and branch order in xor types.  This is done by converting
effect types into a normal form which includes set semantics for the relevant
parts.