chr.factor.phi: Initial loop predicate inference
The type of a recursive word is set to be an Xor with the CallRecursive
and the base case.  The CallRecursive is expanded into an Iteration constraint
during compositional reasoning, and LoopVar relations derived from that.  The
Iteration constraint holds the stacks on entry, last iteration start, and end.

Next:
- Change Iteration constraint to hold stacks on entry, n-th iteration,
n+1-th iteration, and end, otherwise we assume too much when stuff is known
about the end case.
- Make sure to correctly handle more loops with more conditionals in the body
and termination test.