chr.factor: Improve numerical inference, be more strict with liveness
Infer counters in loops, and corresponding relations as soon as the loop step is
known.  This allows to infer e.g. constraints about loop quotation inputs, even
if the quotation is not known, but the bounds are.

Per default, Effect body predicates are now considered live only if all unknowns
are live.  Special predicates, like macro-expansion or Iterated loop markers can
extend this set.

For the case of unknown intermediate arithmetic results, an explicit rule set to
determine which variables are live is added.  This allows defining parameters
depending on the liveness of other parameters of a set of predicates directly.