chr.factor: Somewhat reasonable mode inference
This is the first version which does not take forever
and so far covers all cases.  Basic approach:

1. MakeEffect defines a Def and a Use set
2. Some preds, like Sum, Slot, can extend the Def set,
   and place an implication in the store.
3. If the implication reaches the Use set, then the defining values
   are live and extend the Use set.

CallEffects are always assumed to be live predicates, and their input
stacks are added to the Use set, and the output stack to the Def set.

Iterated constructs which are not resolved yet define a scope for their
loop-body preds, so as to not lose them.

One critical "discovery" was the fact that actually relation predicates like
Eq, Le, etc. don't actually add into modifying the Def and Use sets, but are
only collected based on what the definer predicates and scoping predicates
determine to be live.

Next:
- evaluate performance, this got a bit slower again
- fix the remaining error regarding pushing wrappers
- if the new liveness reasoning approach is correct, clean up
  that code