chr.factor: Use same collect mechanism for phi reasoning like for composition
In other news:
It looks like a FMC-like location reasoning is needed for memory locations.
For the retain stack, hiding it as bi-variable resolution seems to work well
because of the way Factor uses the retain stack, it will always be balanced by
the end of a quotation.  For memory locations, this is not the case.