chr.factor: Properly Support Retain stack operations
Uses a PolyEffect predicate, which represents FMC's poly lambda calculus
sequential compositions with different locations.  However, here the default
assumption is that sequencing can not be changed unless it is known to be safe,
whereas in FMC, all locations are assumed to be independent.  This is because we
actually want to reason about first-class locations, where effects can only
safely be composed if they are known not to alias.

For the retain stack, the assumption is that it is independent of any other
location, and thus effect composition on that location can always proceed.

This constructs kind of inhibits regular effect composition by introducing
bi-variable effects on the *-local primitives.  These are used to enforce
correct sequencing.  Once the RetainEffect predicates are composed such that
only one with an empty effect and empty retain stack remains, the lambda stacks
are unified.

Not completely sure how this resolution condition transfers over to memory locations...