chr.factor: Include FMC fragment
- Sequencing works, but interaction with the existing slot and element
  predicates is tedious.
- Approach: change stuff like the { Slot o n v } pred to be pivoted
  on the location identifier instead of the object.

Right now, successive read-accesses are forced to be in-order.  This is only necessary
for stream location access, not for memory or Retain stack access.  For those
cases, it should be sufficient if the write accesses are correctly ordered
with respect to other write and read accesses.
Thus next: Introduce stateful location content predicate which can be used
for non-destructive memory loc content assertion without force reads.
This should be a safe optimization except for anything that would have to be
considered volatile because of using the system as an adapter for memory mapped
stateful peripheral interfaces, for example.