chr.factor.util: Algorithm for structural matching
This is intended for the unification step during phi reasoning.
Two effects are matched in terms of their data flow structure if no conflicts
or ambiguities arise.  Semantic unification based on the corresponding
predicates can then continue to either compute a union type, or prove that there
are observable differences, as is already implemented.

The implementation is similar to the isomporphism checker, but based on Baader's
UnifyClosure, and extended for bfs for set matching, supporting a very
restricted type of E-unification.